spark/requirements.txt
google-cloud-bigquery
google-cloud-storage
dotenv
datetime
pyarrow
pandas
gcsfs
google-auth==2.41.0